rename to radio_configuration
